# Store AST directly in provider mappings