@ycombinator 2024-11-21
> the CI failures are due to `TestOtelAPMIngestion` (#4148) and `TestRpmFleetUpgrade`. The last one fails because it doesn't download in time the `elastic-agent-8.17.0-SNAPSHOT-x86_64.rpm` artifact with a `context deadline exceeded` ([code ref](https://github.com/elastic/elastic-agent/blob/main/pkg/testing/fetcher_artifact.go#L200-L202))
> 
> cc @ycombinator should we create an issue about the latter?

I'm not seeing these tests failing on `main` so either something in this PR is causing them to fail (unlikely) or they are flaky. 
I just re-ran the integration tests step.  If it passes, it means those tests are flaky and we should [create an issue](https://github.com/elastic/elastic-agent/issues/new?template=flaky-test.yml) for each of them.